Refactor the experimental I2C blocks to a separate module. Remove the I2C module from webpack builds.
