+++ src/modules/toolbox_data.js
@@ -854,33 +854,35 @@ xmlToolbox += '            <block type="serial_status"></block>';
 xmlToolbox += '            <block type="serial_print_multiple"></block>';
 xmlToolbox += '            <block type="serial_receive_text"></block>';
 xmlToolbox += '            <block type="serial_scan_multiple"></block>';
-xmlToolbox += '            <block type="i2c_send" experimental="true">';
-xmlToolbox += '                <value name="DATA">';
-xmlToolbox += '                    <block type="math_number">';
-xmlToolbox += '                        <field name="NUM">10</field>';
-xmlToolbox += '                    </block>';
-xmlToolbox += '                </value>';
-xmlToolbox += '                <value name="ADDR">';
-xmlToolbox += '                    <block type="number_hex"></block>';
-xmlToolbox += '                </value>';
-xmlToolbox += '                <value name="DEVICE">';
-xmlToolbox += '                    <block type="number_hex"></block>';
-xmlToolbox += '                </value>';
-xmlToolbox += '            </block>';
-xmlToolbox += '            <block type="i2c_receive" experimental="true">';
-xmlToolbox += '                <value name="ADDR">';
-xmlToolbox += '                    <block type="number_hex"></block>';
-xmlToolbox += '                </value>';
-xmlToolbox += '                <value name="DEVICE">';
-xmlToolbox += '                    <block type="number_hex"></block>';
-xmlToolbox += '                </value>';
-xmlToolbox += '            </block>';
-xmlToolbox += '            <block type="i2c_busy" experimental="true">';
-xmlToolbox += '                <value name="DEVICE">';
-xmlToolbox += '                    <block type="number_hex"></block>';
-xmlToolbox += '                </value>';
-xmlToolbox += '            </block>';
-xmlToolbox += '            <block type="i2c_mode" experimental="true"></block>';
+
+// xmlToolbox += '            <block type="i2c_send" experimental="true">';
+// xmlToolbox += '                <value name="DATA">';
+// xmlToolbox += '                    <block type="math_number">';
+// xmlToolbox += '                        <field name="NUM">10</field>';
+// xmlToolbox += '                    </block>';
+// xmlToolbox += '                </value>';
+// xmlToolbox += '                <value name="ADDR">';
+// xmlToolbox += '                    <block type="number_hex"></block>';
+// xmlToolbox += '                </value>';
+// xmlToolbox += '                <value name="DEVICE">';
+// xmlToolbox += '                    <block type="number_hex"></block>';
+// xmlToolbox += '                </value>';
+// xmlToolbox += '            </block>';
+// xmlToolbox += '            <block type="i2c_receive" experimental="true">';
+// xmlToolbox += '                <value name="ADDR">';
+// xmlToolbox += '                    <block type="number_hex"></block>';
+// xmlToolbox += '                </value>';
+// xmlToolbox += '                <value name="DEVICE">';
+// xmlToolbox += '                    <block type="number_hex"></block>';
+// xmlToolbox += '                </value>';
+// xmlToolbox += '            </block>';
+// xmlToolbox += '            <block type="i2c_busy" experimental="true">';
+// xmlToolbox += '                <value name="DEVICE">';
+// xmlToolbox += '                    <block type="number_hex"></block>';
+// xmlToolbox += '                </value>';
+// xmlToolbox += '            </block>';
+// xmlToolbox += '            <block type="i2c_mode" experimental="true"></block>';
+
 xmlToolbox += '            <block type="shift_in"></block>';
 xmlToolbox += '            <block type="shift_out">';
 xmlToolbox += '                <value name="VALUE">';
