+++ webpack/dev.config.js
@@ -115,6 +115,7 @@ module.exports = (opts) => {
         {
           test: /\.js$/,
           exclude: [
+            path.resolve(__dirname, './src/modules/blockly/generators/propc/comms/i2c_protocol.js'),
             path.resolve(__dirname, './src/modules/blockly/generators/propc/comms/lcd_parallel.js'),
             path.resolve(__dirname, './src/modules/blockly/generators/propc/comms/wx_simple.js'),
           ]
